% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_defillama_api.R
\name{call_defillama_api}
\alias{call_defillama_api}
\title{Call DefiLlamam API}
\usage{
call_defillama_api(path)
}
\arguments{
\item{path}{The path to pass to the API URL}
}
\value{
a parsed list from the JSON structure
}
\description{
This is the core function connecting to DefiLlama. Since all (most?) API
 calls and authentication this is a very straightforward function.
 It does some standardization of the output for individual handler functions
 to parse
}
\examples{
\dontrun{
# Need a valid API to run
x <- call_defi_api()
}
}
