% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chain_tvl.R
\name{get_chain_tvl}
\alias{get_chain_tvl}
\title{Return the historical TVL per chain}
\usage{
get_chain_tvl(chain = "Ethereum")
}
\arguments{
\item{chain}{a string containing the chain to return}
}
\value{
a tibble with the date and the TVL in USD
}
\description{
Return the historical TVL per chain
}
