% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_tvl.R
\name{get_historical_tvl}
\alias{get_historical_tvl}
\title{Get the historical TVL for all protocols denominated in USD}
\usage{
get_historical_tvl()
}
\value{
a tibble with the date and value
}
\description{
This function returns the sum of all TVLs for all dates available
}
\examples{
x <- get_historical_tvl()
}
