% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protocol_list.R
\name{get_protocol_list}
\alias{get_protocol_list}
\title{Grab the list of protocols from DeFiLlama}
\usage{
get_protocol_list(tvl_limit = 0, which_chain = NULL, details = FALSE)
}
\arguments{
\item{tvl_limit}{This is the minimum level of Total Value Locked to return}

\item{which_chain}{This is what chains to focus on. For ones on multiple chains th
value "Multi-Chain" chould be used}

\item{details}{Should the type and number and list of chains as well as the category
be returned?}
}
\value{
A tibble with date, tvl, chain, number of chains, list of chains, category
}
\description{
This function gets the current list or protocols from DeFiLlama. It can also
returns the number of chains the protocol is on as well as its type. Basic
filter by chain and minimum TVL is also supported
}
\examples{
x <- get_protocol_list(tvl_limit=1E9,which_chain="Polygon")
}
