% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protocol_tvl.R
\name{get_protocol_tvl}
\alias{get_protocol_tvl}
\title{Get a protocol's historic TVL per chain and category}
\usage{
get_protocol_tvl(protocol = "MakerDAO", category = FALSE)
}
\arguments{
\item{protocol}{a string for protocol to return}

\item{category}{a logical whether to return type as a column}
}
\value{
a wide tibble of date, category and one column per chain
}
\description{
This function returns a wide tibble with the historic TVL per
chain of a protocol. It can also return the category of the protocol.
}
\examples{
x <- get_protocol_tvl("WBTC")
}
