% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{GetBlockAtTimestamp}
\alias{GetBlockAtTimestamp}
\title{Get Block Number at Timestamp}
\usage{
GetBlockAtTimestamp(chain, timestamp)
}
\arguments{
\item{chain}{A character string representing the blockchain name (e.g., "ethereum", "polygon").}

\item{timestamp}{A POSIXct object representing the timestamp.}
}
\value{
A tibble containing the block height and timestamp.
}
\description{
This function calls the DefiLlama API and returns the block number for a specific chain at a given timestamp.
}
\examples{
GetBlockAtTimestamp("ethereum", as.POSIXct("2024-01-01"))
}