% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{GetDexVolumeByChain}
\alias{GetDexVolumeByChain}
\title{Get DEX Volume for Specific Chain}
\usage{
GetDexVolumeByChain(chain, excludeTotalDataChart = FALSE, excludeTotalDataChartBreakdown = FALSE)
}
\arguments{
\item{chain}{A character string representing the blockchain name (e.g., "Ethereum", "Arbitrum").}

\item{excludeTotalDataChart}{A logical value indicating whether to exclude chart data. Default is FALSE.}

\item{excludeTotalDataChartBreakdown}{A logical value indicating whether to exclude breakdown data. Default is FALSE.}
}
\value{
A tibble containing DEX volume data for the specified chain.
}
\description{
This function returns DEX volume data for a specific blockchain.
}
\examples{
GetDexVolumeByChain("Ethereum")
}