% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{GetDexVolumeOverview}
\alias{GetDexVolumeOverview}
\title{Get DEX Volume Overview}
\usage{
GetDexVolumeOverview(excludeTotalDataChart = FALSE, excludeTotalDataChartBreakdown = FALSE, dataType = "dailyVolume")
}
\arguments{
\item{excludeTotalDataChart}{A logical value indicating whether to exclude chart data. Default is FALSE.}

\item{excludeTotalDataChartBreakdown}{A logical value indicating whether to exclude breakdown data. Default is FALSE.}

\item{dataType}{A character string specifying data type ("dailyVolume" or "totalVolume"). Default is "dailyVolume".}
}
\value{
A tibble containing DEX volume overview data.
}
\description{
This function returns aggregated DEX volume data across all protocols.
}
\examples{
GetDexVolumeOverview()
}