% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fees.R
\name{GetFeesOverview}
\alias{GetFeesOverview}
\title{Get Fees Overview}
\usage{
GetFeesOverview(excludeTotalDataChart = FALSE, excludeTotalDataChartBreakdown = FALSE, dataType = "dailyFees")
}
\arguments{
\item{excludeTotalDataChart}{A logical value indicating whether to exclude chart data. Default is FALSE.}

\item{excludeTotalDataChartBreakdown}{A logical value indicating whether to exclude breakdown data. Default is FALSE.}

\item{dataType}{A character string specifying data type ("dailyFees", "dailyRevenue", or "dailyHoldersRevenue"). Default is "dailyFees".}
}
\value{
A list containing fees overview data.
}
\description{
This function returns aggregated protocol fees and revenue data.
}
\examples{
GetFeesOverview()
}