% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{GetHistoricalCoinPrices}
\alias{GetHistoricalCoinPrices}
\title{Get Historical Coin Prices}
\usage{
GetHistoricalCoinPrices(coins, timestamp, searchWidth = NULL)
}
\arguments{
\item{coins}{A character vector of coin identifiers (e.g., c("ethereum", "bitcoin")).}

\item{timestamp}{A POSIXct object representing the timestamp for historical prices.}

\item{searchWidth}{A character string representing the time range to search for price ("4h", "24h").}
}
\value{
A tibble containing historical prices for the specified coins.
}
\description{
This function calls the DefiLlama API and returns historical prices for specified coins at a specific timestamp.
}
\examples{
GetHistoricalCoinPrices(c("bitcoin", "ethereum"), as.POSIXct("2024-01-01"))
}