% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{GetOptionsVolumeOverview}
\alias{GetOptionsVolumeOverview}
\title{Get Options Volume Overview}
\usage{
GetOptionsVolumeOverview(excludeTotalDataChart = FALSE, excludeTotalDataChartBreakdown = FALSE)
}
\arguments{
\item{excludeTotalDataChart}{A logical value indicating whether to exclude chart data. Default is FALSE.}

\item{excludeTotalDataChartBreakdown}{A logical value indicating whether to exclude breakdown data. Default is FALSE.}
}
\value{
A tibble containing options volume overview data.
}
\description{
This function returns aggregated options trading volume data.
}
\examples{
GetOptionsVolumeOverview()
}